chore: no proxy, PluginCallbacks
