+++ src/sdk/index.ts
@@ -1,6 +1,6 @@
 export { createPlugin } from "./server";
 export { createActionsPlugin } from "./actions";
-export { CallbackBuilder } from "./proxy-callbacks";
+export { CallbackBuilder } from "./plugin-callbacks";
 export { postComment } from "./comment";
 export type { Context } from "./context";
 export * from "./constants";
